+++ src/board/v3/nap/nap_common.c
@@ -259,8 +259,9 @@ void nap_pps(u32 count) {
 }
 
 void nap_pps_config(u32 microseconds, u8 active) {
-  u32 width =
-      ceil((double)microseconds / ((1.0 / NAP_TIMING_COUNT_RATE_Hz) * 1e6)) - 1;
+  u32 width = (u32)ceil((double)microseconds /
+                        ((1.0 / NAP_TIMING_COUNT_RATE_Hz) * 1e6)) -
+              1;
   NAP->PPS_CONTROL =
       (width << NAP_PPS_CONTROL_PULSE_WIDTH_Pos) | (active & 0x01);
 }
@@ -289,7 +290,8 @@ u32 nap_get_ext_event(u8 pin, ext_event_trigger_t *trig) {
 }
 
 void nap_set_ext_event(u8 pin, ext_event_trigger_t trig, u32 timeout) {
-  u32 gap = ceil((double)timeout / ((1.0 / NAP_TIMING_COUNT_RATE_Hz) * 1e6));
+  u32 gap =
+      (u32)ceil((double)timeout / ((1.0 / NAP_TIMING_COUNT_RATE_Hz) * 1e6));
 
   switch (pin) {
     case 0:
